Add files via upload
This is an archive with two bash scripts, which can be used as the tutorials of the molecular dynamics simulations of the lipid bilayer systems with different physical conditions. In this examples you can perform molecular modeling of POPC lipid bilayer at different temperatures (burn_bilayer.sh) and/or under different pressure (compress_bilayer.sh) in order to investigate physical behaviour of the biomembrane. Depending on your scientific creativity you can easily vary additional physical parameters of the simulations (e.g chosing different barostats, thermostats, introduction of the forces to the parts of the system) or try to adapt this scripts for your molecular systems.
It can be used as the excellent introduction to the molecular dynamics simulations of GROMACS as well as the tutorials of the shell scripting applied on the molecular modeling and computational biophysics.